+++ src/validation.h
@@ -47,7 +47,6 @@
 #include <vector>
 
 class Chainstate;
-class CBlockTreeDB;
 class CTxMemPool;
 class ChainstateManager;
 struct ChainTxData;
